requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper

# For FFmpeg dependency 
# Note: FFmpeg itself must be installed on the server system